split Forms from inout
